requirements.txt

Flask==1.1.1
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
pymessenger==0.0.7.0
python-decouple==3.3
psycopg2-binary==2.8.4